// src/app/models/TransportationModels/TransCalReq.model.ts
export class TransCalReq {
	minM2: number;
	maxM2: number;
	fromFloor: number;
	toFloor: number;
	fromElevatorType: number;
	toElevatorType: number;
	fromWalkingWay: number;
	toWalkingWay: number;
	numberOfPerson: number;
	fromLatitude: number;
	fromLongitude: number;
	toLatitude: number;
	toLongitude: number;
	dinstanceInKM: number;
	estimatedPrice: boolean;
}